Moving LTL to a new package. Separating LTL from BoundedLTL.
