+++ src/main/java/eu/iv4xr/framework/LTL/BoundedLTL.java
@@ -1,4 +1,4 @@
-package eu.iv4xr.framework.MainConcepts;
+package eu.iv4xr.framework.LTL;
 
 import java.util.*;
 import java.util.function.Function;
@@ -147,9 +147,9 @@ public class BoundedLTL implements EnvironmentInstrumenter {
 	 * @param env
 	 * @return
 	 */
-	VERDICT sat(Environment env) {
+	LTLVerdict sat(Environment env) {
 		switch(bltlState) {
-			case SATFOUND : return VERDICT.SAT ;
+			case SATFOUND : return LTLVerdict.SAT ;
 			case NOTSTARTED :
 				if (startf.test(env)) {
 					registerToTrace(env) ;
@@ -159,7 +159,7 @@ public class BoundedLTL implements EnvironmentInstrumenter {
 					if (endf.test(env)) {
 						// the interval ends immediately
 						var verdict = ltl.sat() ;
-						if (verdict == VERDICT.SAT) {
+						if (verdict == LTLVerdict.SAT) {
 							bltlState = BLTLstate.SATFOUND ;
 						}
 						else {
@@ -168,7 +168,7 @@ public class BoundedLTL implements EnvironmentInstrumenter {
 						return verdict ;
 					}
 				}
-				return VERDICT.UNKNOWN ;
+				return LTLVerdict.UNKNOWN ;
 			case STARTED :
 				// pass the env to the ltl to have its atoms evaluated:
 				registerToTrace(env) ;
@@ -177,7 +177,7 @@ public class BoundedLTL implements EnvironmentInstrumenter {
 				if (endf.test(env)) {
 					// end marker holds; then force full evaluation of the ltl
 					var verdict = ltl.sat() ;
-					if (verdict == VERDICT.SAT) {
+					if (verdict == LTLVerdict.SAT) {
 						bltlState = BLTLstate.SATFOUND ;
 					}
 					else {
@@ -192,16 +192,16 @@ public class BoundedLTL implements EnvironmentInstrumenter {
 						bltlState = BLTLstate.NOTSTARTED ;
 						trace.reset();
 					}
-					return VERDICT.UNKNOWN ;
+					return LTLVerdict.UNKNOWN ;
 				}
 		}
 		// should not reach this point
 		return null ;
 	}
 	
-	public VERDICT getVerdict() {
-		if (bltlState == BLTLstate.SATFOUND) return VERDICT.SAT ;
-		return VERDICT.UNSAT ;
+	public LTLVerdict getVerdict() {
+		if (bltlState == BLTLstate.SATFOUND) return LTLVerdict.SAT ;
+		return LTLVerdict.UNSAT ;
 	}
 	
 	public ExecutionTrace getWitness() {
@@ -209,267 +209,10 @@ public class BoundedLTL implements EnvironmentInstrumenter {
 		return null ;
 	}
 
-	/**
-	 * Representing an LTL formula.
-	 */
-	public abstract static class LTL {
-		
-		/**
-		 * Abstractlty representing the finite execution on which this LTL is interpreted.
-		 */
-		LinkedList<VerdictInfo> absexecution = new LinkedList<VerdictInfo>() ;
-		
-		LTL() { }
-		
-		
-		abstract void resettracking() ;
-		
-		/**
-		 * Check if {@link #absexecution} satisfies this LTL formula.
-		 */
-		abstract VERDICT sat() ;
-		
-		/**
-		 * Evaluate the atoms of this LTL formula on the given env (concrete state),
-		 * and add the verdict at the end of {@link #absexecution}.
-		 */
-		abstract void evalAtomSat(Environment env) ;	
-
-		/**
-		 * Construct the LTL formula "phi Until psi", where phi is this LTL formula.
-		 */
-		public LTL ltlUntil(LTL psi) {
-			var ltl = new Until() ;
-			ltl.phi1 = this ;
-			ltl.phi2 = psi ;
-			return ltl ;
-		}
+	public enum LTLVerdict { SAT, UNSAT, UNKNOWN }
+	public static class LTLVerdictInfo { 
+		public LTLVerdict verdict ;
+		LTLVerdictInfo(LTLVerdict v) { verdict = v ; }
 	}
 	
-	public enum VERDICT { SAT, UNSAT, UNKNOWN }
-	public static class VerdictInfo { 
-		public VERDICT verdict ;
-		VerdictInfo(VERDICT v) { verdict = v ; }
-	}
-	
-	
-	public static class Atom extends LTL {
-		Predicate<Environment> p ;
-		void check(Environment env) {
-					
- 		}
-		@Override
-		void resettracking() { absexecution.clear(); }
-		
-		@Override
-		VERDICT sat() {
-			return absexecution.getFirst().verdict ;
-		}
-		
-		@Override
-		void evalAtomSat(Environment env) {
-			if (p.test(env)) 
-				absexecution.add(new VerdictInfo(VERDICT.SAT)) ;
-			else
-				absexecution.add(new VerdictInfo(VERDICT.UNSAT)) ;		
-		}
-	}
-	
-	public static class Not extends LTL {
-		LTL phi ;
-		
-		@Override
-		void resettracking() {
-			absexecution.clear(); phi.resettracking();	
-		}
-
-		@Override
-		VERDICT sat() {
-			phi.sat() ;
-			var iterator = absexecution.descendingIterator() ;
-			var iteratorPhi = phi.absexecution.descendingIterator() ;
-			
-			while (iterator.hasNext()) {
-				var psi = iterator.next() ;
-				var p = iteratorPhi.next().verdict ;
-				switch (p) {
-				case SAT : psi.verdict = VERDICT.UNSAT ; break ;
-				case UNSAT : psi.verdict = VERDICT.SAT ; break ;
-				}
-			}
-			
-			return absexecution.getFirst().verdict ;
-		}
-
-		@Override
-		void evalAtomSat(Environment env) {
-			absexecution.add(new VerdictInfo(VERDICT.UNKNOWN)) ;
-			phi.evalAtomSat(env);		
-		}
-	}
-	
-	public static class And extends LTL {
-		LTL[] conjuncts ;
-		
-		@Override
-		void resettracking() {
-			absexecution.clear(); 
-			for (LTL phi : conjuncts) phi.resettracking();	
-		}
-		
-		@Override
-		VERDICT sat() {
-			for (LTL phi : conjuncts) phi.sat() ;
-			var iterator = absexecution.descendingIterator() ;
-			var N = conjuncts.length ;
-			Iterator<VerdictInfo>[] conjuntIterators = new Iterator[N] ;
-			for (int k=0; k<N; k++)
-				conjuntIterators[k] = conjuncts[k].absexecution.descendingIterator() ;
-		
-			while (iterator.hasNext()) {
-				var psi = iterator.next() ;
-				boolean allSat = true ;
-				for (int k=0; k<N; k++) {
-					var p = conjuntIterators[k].next().verdict ;
-					allSat = allSat && (p == VERDICT.SAT) ;
-				}
-				if (allSat) psi.verdict = VERDICT.SAT ;
-				else psi.verdict = VERDICT.UNSAT ;
-			}
-			return absexecution.getFirst().verdict ;
-		}
-		
-		@Override
-		void evalAtomSat(Environment env) {
-			absexecution.add(new VerdictInfo(VERDICT.UNKNOWN)) ;
-			for (LTL phi : conjuncts) phi.evalAtomSat(env);		
-		}
-	}
-	
-	
-	public static class Until extends LTL {
-		LTL phi1 ;
-		LTL phi2 ;
-		
-		@Override
-		void resettracking() {
-			absexecution.clear(); 
-			phi1.resettracking();	
-			phi2.resettracking();	
-		}
-		
-		@Override
-		VERDICT sat() {
-			phi1.sat() ;
-			phi2.sat() ;
-			var iterator = absexecution.descendingIterator() ;
-			var iteratorPhi1 = phi1.absexecution.descendingIterator() ;
-			var iteratorPhi2 = phi2.absexecution.descendingIterator() ;
-			
-			// keep track if phi1 untill phi2 holds at sigma(k+1)
-			boolean nextSat = false ;
-			
-			// calculate phi1 until phi2 holds on every sigma(k); we calculate this
-			// backwards for every state in the interval:
-			while (iterator.hasNext()) {
-				var psi = iterator.next() ;
-				var p = iteratorPhi1.next().verdict ;
-				var q = iteratorPhi2.next().verdict ;
-				if (q == VERDICT.SAT) {
-					psi.verdict = VERDICT.SAT ;
-					nextSat = true ;
-				}
-				else {
-					if (nextSat && p == VERDICT.SAT) psi.verdict = VERDICT.SAT ;
-					else {
-						psi.verdict = VERDICT.UNSAT ;
-						nextSat = false ;	
-					}
-				}
-			}
-			return absexecution.getFirst().verdict ;
-		}
-		
-		@Override
-		void evalAtomSat(Environment env) {
-			absexecution.add(new VerdictInfo(VERDICT.UNKNOWN)) ;
-			phi1.evalAtomSat(env);	
-			phi2.evalAtomSat(env);			
-		}						
-	}
-	
-	static public class Next extends LTL {
-		
-		LTL phi ;
-
-		@Override
-		void resettracking() {
-			absexecution.clear(); 
-			phi.resettracking();		
-		}
-
-		@Override
-		VERDICT sat() {
-			phi.sat() ;
-			var iterator = absexecution.descendingIterator() ;
-			var iteratorPhi = phi.absexecution.descendingIterator() ;
-			
-			var psi = iterator.next() ;
-			psi.verdict = VERDICT.UNSAT ; // always unsat at the last state
-
-			// calculate phi1 until phi2 holds on every sigma(k); we calculate this
-			// backwards for every state in the interval:
-			while (iterator.hasNext()) {
-				psi = iterator.next() ;
-				var q = iteratorPhi.next().verdict ;
-				switch(q) {
-				  case SAT   : psi.verdict = VERDICT.SAT ; break ;
-				  case UNSAT : psi.verdict = VERDICT.UNSAT ;
-				}
-			}
-			
-			return absexecution.getFirst().verdict ;
-		}
-
-		@Override
-		void evalAtomSat(Environment env) {
-			absexecution.add(new VerdictInfo(VERDICT.UNKNOWN)) ;
-			phi.evalAtomSat(env);			
-		}
-		
-	}
-	
-	public static <E> LTL now(Predicate<E> p) {
-		var a = new Atom() ;
-		a.p = env -> p.test((E) env) ;
-		return a ;
-	}
-	
-	public static LTL next(LTL phi) { 
-		var ltl = new Next() ;
-		ltl.phi = phi ;
-		return ltl ;
-	}
-	
-	public static LTL ltlNot(LTL phi) {
-		var ltl = new Not() ;
-		ltl.phi = phi ; return ltl ;
-	}
-	
-	public static LTL ltlAnd(LTL... phis) {
-		if (phis == null) throw new IllegalArgumentException() ;
-		if (phis.length <2) throw new IllegalArgumentException() ;
-		var ltl = new And() ;
-		ltl.conjuncts = phis ;
-		return ltl ;
-	}
-	
-	public static LTL eventually(LTL phi) {
-		return now((Environment env) -> true).ltlUntil(phi) ;
-	}
-	
-	public static LTL always(LTL phi) {
-		return ltlNot(eventually(ltlNot(phi))) ;
-	}
-
 }
